@@ -16,11 +16,11 @@
 
 package com.example.android.architecture.blueprints.todoapp.addedittask;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-
 import com.example.android.architecture.blueprints.todoapp.data.Task;
 import com.example.android.architecture.blueprints.todoapp.data.source.TasksDataSource;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
