@@ -16,9 +16,9 @@
 
 package com.example.android.architecture.blueprints.todoapp.util;
 
-import android.support.test.espresso.IdlingResource;
-
 import java.util.concurrent.atomic.AtomicInteger;
+
+import androidx.test.espresso.IdlingResource;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
